- FragmentMain추가함. - 로그관련코드 수정함.

+++ app/src/main/java/com/kct/iqsdisplayer/network/Packet.kt
@@ -57,7 +57,7 @@ class Packet(buffer: ByteArray) {
             while (data.hasRemaining()) {
                 temp = data.get()
                 if (temp.toInt() == 0x00) break
-                sb.append(temp.toChar())
+                sb.append(temp.toInt().toChar())
             }
             val result = String(sb.toString().toByteArray(charset("euc-kr"))).trim()
             Log.d("getString : $result")
